+++ event.go
@@ -43,7 +43,7 @@ type severity struct {
 }
 
 // The form of stacktrace that Bugsnag expects
-type stackFrame struct {
+type StackFrame struct {
 	Method     string `json:"method"`
 	File       string `json:"file"`
 	LineNumber int    `json:"lineNumber"`
@@ -85,7 +85,7 @@ type Event struct {
 	// The error message to be sent to Bugsnag. This defaults to the return value of Error.Error()
 	Message string
 	// The stacktrrace of the error to be sent to Bugsnag.
-	Stacktrace []stackFrame
+	Stacktrace []StackFrame
 
 	// The context to be sent to Bugsnag. This should be set to the part of the app that was running,
 	// e.g. for http requests, set it to the path.
@@ -123,6 +123,7 @@ func newEvent(rawData []interface{}, notifier *Notifier) (*Event, *Configuration
 	}
 
 	var err *errors.Error
+	var callbacks []func(*Event)
 
 	for _, datum := range event.RawData {
 		switch datum := datum.(type) {
@@ -135,7 +136,7 @@ func newEvent(rawData []interface{}, notifier *Notifier) (*Event, *Configuration
 				event.ErrorClass = err.TypeName()
 			}
 			event.Message = err.Error()
-			event.Stacktrace = make([]stackFrame, len(err.StackFrames()))
+			event.Stacktrace = make([]StackFrame, len(err.StackFrames()))
 
 		case bool:
 			config = config.merge(&Configuration{Synchronous: bool(datum)})
@@ -169,6 +170,8 @@ func newEvent(rawData []interface{}, notifier *Notifier) (*Event, *Configuration
 		case HandledState:
 			event.handledState = datum
 			event.Severity = datum.OriginalSeverity
+		case func(*Event):
+			callbacks = append(callbacks, datum)
 		}
 	}
 
@@ -184,11 +187,18 @@ func newEvent(rawData []interface{}, notifier *Notifier) (*Event, *Configuration
 			file = config.stripProjectPackages(file)
 		}
 
-		event.Stacktrace[i] = stackFrame{
+		event.Stacktrace[i] = StackFrame{
 			Method:     frame.Name,
 			File:       file,
 			LineNumber: frame.LineNumber,
 			InProject:  inProject,
+		}
+	}
+
+	for _, callback := range callbacks {
+		callback(event)
+		if event.Severity != event.handledState.OriginalSeverity {
+			event.handledState.SeverityReason = SeverityReasonCallbackSpecified
 		}
 	}