+++ CHANGELOG.md
@@ -6,6 +6,21 @@
 
 * Extract stacktrace contents on errors wrapped by
   [`pkg/errors`](https://github.com/pkg/errors).
+* Support modifying an individual event using a callback function argument.
+
+  ```go
+  bugsnag.Notify(err, func(event *bugsnag.Event) {
+    event.ErrorClass = "Unexpected Termination"
+    event.MetaData.Update(loadJobData())
+
+    if event.Stacktrace[0].File = "mylogger.go" {
+      event.Stacktrace = event.Stacktrace[1:]
+    }
+  })
+  ```
+
+  The stack trace of an event is now mutable so frames can be removed or
+  modified.
 
 ### Bug fixes